A completely frivolous application for encrypting private messages))
A completely frivolous application for encrypting private messages)) Hey! My name is Maksim Ermak and I wrote this application for my son.
IMPORTANT: do not take it seriously, this is, in general, a joke). The application allows you to write a message,
encrypt it, copy and paste it into any messenger. The recipient using the same application,
installed already on his smartphone, will easily decipher this message).
Do not expect any serious crypto protection, that messages cannot be decrypted - well, not ENIGMA this).
And one more important point: the text must be entered IN LOWER LETTERS - then everything will work correctly.
Once again - treat the application with ease, like a toy, a funny joke - which, in fact, it is.
All good!)))
The app supports English, Spanish and German.